+++ packages/terra-avatar/src/terra-dev-site/test/avatar/facility/DefaultFacility.test.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import { Facility } from '../../../../index';
-
-export default () => <Facility id="default-facility" alt="facility" color="three" />;