+++ packages/terra-core-docs/src/terra-dev-site/doc/avatar/example/facility/FacilityImage.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { Facility } from 'terra-avatar';
+import image from '../../assets/200x133.jpg';
+
+const FacilityImage = () => (
+  <Facility image={image} alt="Thailand" />
+);
+
+export default FacilityImage;