+++ packages/terra-core-docs/src/terra-dev-site/test/avatar/avatar/OneInitialAvatar.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import Avatar from 'terra-avatar';
+
+export default () => <Avatar id="one-initial-avatar" initials="J" alt="John" />;