+++ packages/terra-core-docs/src/terra-dev-site/test/avatar/facility/DefaultFacility.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import { Facility } from 'terra-avatar';
+
+export default () => <Facility id="default-facility" alt="facility" color="three" />;